requirements.txt
Flask==2.0.1
gunicorn==20.1.0
tensorflow-cpu==2.8.0
numpy==1.21.2
Pillow==8.3.2
scikit-image==0.18.2
fpdf==1.7.2
python-dotenv==0.19.0
opencv-python-headless==4.5.3.56
scikit-learn==0.24.2
shap==0.41.0
lime==0.2.0.1
matplotlib==3.5.1
seaborn==0.11.2 